src/main/java/com/hr/recruitment/api/dto/ApplicationDTO.java
package com.hr.recruitment.api.dto;

import java.io.Serializable;

public class ApplicationDTO implements Serializable{

	private static final long serialVersionUID = -628560387152469121L;

	private Long applicationId;
	
	private Long offerId;
	
	private String candidateEmail;
	
	private String resumeText;

	private String status;

	public Long getApplicationId() {
		return applicationId;
	}

	public void setApplicationId(Long applicationId) {
		this.applicationId = applicationId;
	}

	public Long getOfferId() {
		return offerId;
	}

	public void setOfferId(Long offerId) {
		this.offerId = offerId;
	}

	public String getCandidateEmail() {
		return candidateEmail;
	}

	public void setCandidateEmail(String candidateEmail) {
		this.candidateEmail = candidateEmail;
	}

	public String getResumeText() {
		return resumeText;
	}

	public void setResumeText(String resumeText) {
		this.resumeText = resumeText;
	}

	public String getStatus() {
		return status;
	}

	public void setStatus(String status) {
		this.status = status;
	}
	
}
